+++ lib/curves/base.js
