+++ lib/utils/calculateHeight.js
