+++ lib/utils/constructSource.js
