+++ lib/utils/crypto.js
