+++ lib/utils/datatypes.js
