+++ lib/utils/errors.js
