+++ lib/v1/core/crypto.js
@@ -1,7 +1,7 @@
 /** @format */
 
-const { Datatypes, errors, cryptoUtils } = require("../utils");
-const { arrayBufferToBase64 } = require("../utils/datatypes");
+const { Datatypes, errors, cryptoUtils } = require("../../utils");
+const { arrayBufferToBase64 } = require("../../utils/datatypes");
 
 const generateBytes = (byteLength) => {
   return new Promise((resolve, reject) => {
