+++ lib/v1/core/http.js
@@ -1,4 +1,4 @@
-const { errors } = require("../utils");
+const { errors } = require("../../utils");
 
 module.exports = (config, teamId, context) => {
   if (!window.fetch)
