+++ lib/v1/core/inputs.js
@@ -1,7 +1,7 @@
 /** @format */
 
-const constructSource = require("../utils/constructSource");
-const calculateHeight = require("../utils/calculateHeight");
+const constructSource = require("../../utils/constructSource");
+const calculateHeight = require("../../utils/calculateHeight");
 
 module.exports = (config) => {
   return {
