+++ lib/v2/core/http.js
@@ -1,4 +1,4 @@
-const { errors } = require("../utils");
+const { errors } = require("../../utils");
 
 module.exports = (config, teamId, appId, context) => {
   if (!window.fetch)
