--- lib/v2/curves/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  P256: require("./p256"),
-};
