+++ lib/v2/index.js
@@ -1,7 +1,7 @@
-const { Datatypes, errors, extractDomain } = require("./utils");
+const { Datatypes, errors, extractDomain } = require("../utils");
 const Config = require("./config");
 const { Crypto, Http, Forms, Input } = require("./core");
-const { deriveSharedSecret } = require("./utils/crypto");
+const { deriveSharedSecret } = require("../utils/crypto");
 
 class EvervaultClient {
   constructor(teamId, appId, customConfig = {}) {
