--- lib/v2/utils/calculateHeight.js
@@ -1,7 +1,0 @@
-module.exports = function (theme, styles) {
-  if (styles && styles.height) {
-    return styles.height;
-  }
-  if (theme === "minimal" || theme === "material") return "145px";
-  return "180px";
-};
