--- lib/v2/utils/domain.js
@@ -1,18 +1,0 @@
-const extractDomainFromUrl = (url) => {
-  var result;
-  var match;
-  url = url.trim();
-  if (
-    (match = url.match(
-      /^(?:https?:\/\/)?(?:[^@\n]+@)?(?:www\.)?([^:\/\n\?\=]+)/im
-    ))
-  ) {
-    result = match[1];
-    if ((match = result.match(/^[^\.]+\.(.+\..+)$/))) {
-      result = match[1];
-    }
-  }
-  return result;
-};
-
-module.exports = extractDomainFromUrl;
