--- lib/v2/utils/index.js
@@ -1,7 +1,0 @@
-module.exports = {
-  Datatypes: require("./datatypes"),
-  errors: require("./errors"),
-  uuid: require("./uuid"),
-  extractDomain: require("./domain"),
-  cryptoUtils: require("./crypto"),
-};
